fix: remove deleted change on reparse
