+++ packages/eslint-plugin-fiori-tools/src/project-context/parser/parser.ts
@@ -23,7 +23,8 @@ import type {
     FlexChange
 } from './types.js';
 import { uniformUrl } from '@sap-ux/fiori-annotation-api';
-import { existsSync, readdirSync, readFileSync } from 'node:fs';
+import { existsSync, readdirSync } from 'node:fs';
+import { collectFlexChanges } from '../utils.js';
 
 export interface ParseResult {
     index: ParsedProject;
@@ -83,22 +84,11 @@ export class ApplicationParser {
                 const [parsedManifest, services] = this.parseManifest(webappPath, manifestUri, manifest);
                 const appRootUri = pathToFileURL(app.appRoot).toString();
                 const changes: FlexChange[] = [];
-                if (existsSync(join(app.appRoot, 'webapp', 'changes'))) {
-                    const changeFiles = readdirSync(join(app.appRoot, 'webapp', 'changes'))
+                if (existsSync(join(webappPath, 'changes'))) {
+                    const changeFiles = readdirSync(join(webappPath, 'changes'))
                         .filter((file) => file.endsWith('propertyChange.change'))
-                        .map((file) => normalizePath(join(app.appRoot, 'webapp', 'changes', file)));
-                    for (const changeFile of changeFiles) {
-                        const changeFileUri = pathToFileURL(changeFile).toString();
-                        const fileContent =
-                            fileCache.get(changeFileUri) ?? readFileSync(changeFile, { encoding: 'utf8', flag: 'r' });
-                        const jsonContent = JSON.parse(fileContent);
-                        changes.push({
-                            changeType: jsonContent.changeType,
-                            content: jsonContent.content,
-                            selector: jsonContent.selector,
-                            changeFileUri: pathToFileURL(changeFile).toString()
-                        });
-                    }
+                        .map((file) => normalizePath(join(webappPath, 'changes', file)));
+                    changes.push(...collectFlexChanges(changeFiles));
                 }
                 const parsedApp: ParsedApp = {
                     manifest: parsedManifest,
@@ -214,6 +204,14 @@ export class ApplicationParser {
      * @param fileCache - Map of file URIs to their contents
      */
     private reparseChange(uri: string, index: ParsedProject, fileCache: Map<string, string>): void {
+        // Remove deleted files from app changes array
+        for (const key of Object.keys(index.apps)) {
+            const existingChanges = index.apps[key].changes.filter((change) => {
+                const path = fileURLToPath(change.changeFileUri);
+                return existsSync(path);
+            });
+            index.apps[key].changes = existingChanges;
+        }
         const content = fileCache.get(uri) ?? '';
         const ast = parseJson(content, {
             mode: 'json',
@@ -222,7 +220,7 @@ export class ApplicationParser {
             allowTrailingCommas: false
         });
         index.documents[uri] = ast;
-        const jsonContent = JSON.parse(content);
+        const jsonContent = JSON.parse(content) as FlexChange;
         const change: FlexChange = {
             changeType: jsonContent.changeType,
             content: jsonContent.content,
